Improve Add Task and Edit Task feature
In both modals, when toggling to choose what type of payment the client chose, the amount array will become empty.

When it comes to assigning employees for the task, the button to be able to assign employees will be interchanged between add or edit. based on the length of the employees The former is when the length of the employees is equivalent to 0, and the latter is when the length of the employees is greater than 0.

When the user tries to assign employees, each index in the amount array will have an empty string using the clearAmount function.