Task Features
New Task Data:
In Task, name of the task will not be unique. Additional data named priority is added.

New Task Features:
*Edit Task - a feature that can edit data of the selected Task.
*Delete Task - a feature that can delete selected task.
*Show Task - a feature that can show the data of the selected task.

Hourlog Data:
Hourlog changes reference key from taskName and projectId to only taskId. Hourlog is now included date.
isRunning column is removed.
amount column is included in the hourlog.

In accessing data, rows of data are now arranged in ascending order.

In deleting a project, the way to delete hourlog is changed because of the change of reference key.